feat: transaction table actions

+++ app/api/transaction/route.ts
@@ -3,6 +3,7 @@ import { getTransactions } from "@/lib/fetchers/get-transactions";
 import { prisma } from "@/lib/prisma";
 import { IdArraySchema, paginatedQuerySchema } from "@/lib/types/common.types";
 import { baseTransactionSchema } from "@/modules/transaction/transaction-form.validation";
+import { TransactionStatus } from "@/prisma/client/enums";
 import { NextRequest, NextResponse } from "next/server";
 
 export const POST = routeGuard(async (request: NextRequest, { user }) => {
@@ -81,29 +82,45 @@ export const DELETE = routeGuard(async (request: NextRequest, { user }) => {
   const payload = IdArraySchema.safeParse(await request.json());
 
   if (!payload.success) {
-    return NextResponse.json(
-      {
-        message: "Invalid data",
-      },
-      { status: 400 }
-    );
+    return NextResponse.json({ message: "Invalid data" }, { status: 400 });
   }
 
   try {
-    await prisma.transaction.deleteMany({
+    const transactionsToDelete = await prisma.transaction.findMany({
       where: {
         id: { in: payload.data.ids },
         sellerId: user.id,
+      },
+      select: { id: true, status: true },
+    });
+
+    const hasConfirmedTransaction = transactionsToDelete.some(
+      (t) => t.status === TransactionStatus.CONFIRMED
+    );
+
+    if (hasConfirmedTransaction) {
+      return NextResponse.json(
+        { message: "Cannot delete confirmed transactions." },
+        { status: 403 }
+      );
+    }
+
+    const result = await prisma.transaction.deleteMany({
+      where: {
+        id: { in: payload.data.ids },
+        sellerId: user.id,
+        status: TransactionStatus.DRAFT,
       },
     });
 
     return NextResponse.json(
       {
-        message: `${payload.data.ids.length} transaction(s) deleted successfully`,
+        message: `${result.count} transaction(s) deleted successfully`,
       },
       { status: 200 }
     );
   } catch (error) {
+    console.error("Delete Error:", error);
     return NextResponse.json(
       { message: "Internal Server Error" },
       { status: 500 }
